Fix a crash while installing a package
In some cases a file name returned by `git ls-files` or `hg manifest`
could be an empty directory name and if so trying to open it will result
in a crash. This happens for example in the case of a git sub module
directory from which no files are being installed.

Related to nim-lang/nimble#127