Change path command to return all installed paths
Nimble's `path` command is changed to return all installed paths for a
package, but not only the latest version, because multiple latest
versions with different checksums can be installed. The command is no
longer capable of giving paths to develop packages because already there
is no record with a link file for them in the cache.

Related to nim-lang/nimble#127