Fix a bug in the `asyncpipe.createPipe` procedure
In the `asyncpipe.createPipe` procedure the `OVERLAPPED` object passed
to the `connectNamedPipe` and `getOverlappedResult` procedures, was
created on the stack instead of on the heap. It must be created on the
heap as a `PCustomOverlapped` object, the same way as the other places
in the library.

Related to nim-lang/nimble#127