Make special version a set of versions
Special versions are made a set of versions. Those are aliases with
which a single package can be referred. For example, a package can be
simultaneously versions:

 * 0.1.0 - the normal version from the Nimble file.

 * #head - the latest commit in the main branch

 * #master - the main branch name

 * 3c91b869 - part of the sha1 hash of the latest commit in the main
              branch

When the same package is downloaded a second time (determined by the
checksum) instead of proposing to replace it just print a warning that
the package is already installed and merge the special version of the
new package with a special version of the already installed one.

Additionally this commit:

 - Removes some legacy code for supporting the old package format in the
   reverse dependencies.

 - The names of the packages in the reverse dependencies are written
   without converting to lower case.

 - The tests are fixed according to the new behavior.

Related to nim-lang/nimble#127